Refractor packages and activities
To improve the structure of Launcher
