+++ app/src/main/java/com/finnmglas/launcher/choose/ChooseSectionsPagerAdapter.kt
@@ -5,6 +5,8 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
 import com.finnmglas.launcher.*
+import com.finnmglas.launcher.choose.apps.ChooseFragmentApps
+import com.finnmglas.launcher.choose.other.ChooseFragmentOther
 
 private val TAB_TITLES = arrayOf(
     R.string.choose_tab_app,
