+++ app/src/main/java/com/finnmglas/launcher/choose/apps/ChooseFragmentApps.kt
@@ -1,4 +1,4 @@
-package com.finnmglas.launcher.choose
+package com.finnmglas.launcher.choose.apps
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,14 +7,15 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.finnmglas.launcher.R
-import com.finnmglas.launcher.choose.other.OtherRecyclerAdapter
-import com.finnmglas.launcher.extern.dominantColor
-import com.finnmglas.launcher.extern.getSavedTheme
-import kotlinx.android.synthetic.main.fragment_choose_other.*
+import com.finnmglas.launcher.choose.action
+import com.finnmglas.launcher.extern.*
+import com.finnmglas.launcher.choose.forApp
+import kotlinx.android.synthetic.main.fragment_choose_apps.*
 
-/** The 'Other' Tab associated Fragment in the Chooser */
 
-class ChooseFragmentOther : Fragment() {
+/** The 'Apps' Tab associated Fragment in the Chooser */
+
+class ChooseFragmentApps : Fragment() {
 
     /** Lifecycle functions */
 
@@ -22,25 +23,30 @@ class ChooseFragmentOther : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_choose_other, container, false)
+        return inflater.inflate(R.layout.fragment_choose_apps, container, false)
     }
 
     override fun onStart() {
+        super.onStart()
+
         if (getSavedTheme(context!!) == "custom") {
-            fragment_choose_other_container.setBackgroundColor(dominantColor)
+            fragment_choose_apps_container.setBackgroundColor(dominantColor)
         }
 
         // set up the list / recycler
         val viewManager = LinearLayoutManager(context)
-        val viewAdapter = OtherRecyclerAdapter(activity!!)
+        val viewAdapter = AppsRecyclerAdapter(
+            activity!!,
+            action,
+            forApp
+        )
 
-        fragment_choose_other_recycler_view.apply {
+        fragment_choose_apps_recycler_view.apply {
             // improve performance (since content changes don't change the layout size)
             setHasFixedSize(true)
             layoutManager = viewManager
             adapter = viewAdapter
         }
 
-        super.onStart()
     }
 }
