+++ app/src/main/java/com/finnmglas/launcher/choose/other/ChooseFragmentOther.kt
@@ -1,4 +1,4 @@
-package com.finnmglas.launcher.choose
+package com.finnmglas.launcher.choose.other
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,17 +6,14 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.finnmglas.launcher.R
-import com.finnmglas.launcher.action
-import com.finnmglas.launcher.extern.*
-import com.finnmglas.launcher.forApp
-import kotlinx.android.synthetic.main.fragment_choose_apps.*
+import com.finnmglas.launcher.extern.dominantColor
+import com.finnmglas.launcher.extern.getSavedTheme
+import kotlinx.android.synthetic.main.fragment_choose_other.*
 
+/** The 'Other' Tab associated Fragment in the Chooser */
 
-/** The 'Apps' Tab associated Fragment in the Chooser */
-
-class ChooseFragmentApps : Fragment() {
+class ChooseFragmentOther : Fragment() {
 
     /** Lifecycle functions */
 
@@ -24,26 +21,25 @@ class ChooseFragmentApps : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_choose_apps, container, false)
+        return inflater.inflate(R.layout.fragment_choose_other, container, false)
     }
 
     override fun onStart() {
-        super.onStart()
-
         if (getSavedTheme(context!!) == "custom") {
-            fragment_choose_apps_container.setBackgroundColor(dominantColor)
+            fragment_choose_other_container.setBackgroundColor(dominantColor)
         }
 
         // set up the list / recycler
         val viewManager = LinearLayoutManager(context)
-        val viewAdapter = AppsRecyclerAdapter( activity!!, action, forApp)
+        val viewAdapter = OtherRecyclerAdapter(activity!!)
 
-        fragment_choose_apps_recycler_view.apply {
+        fragment_choose_other_recycler_view.apply {
             // improve performance (since content changes don't change the layout size)
             setHasFixedSize(true)
             layoutManager = viewManager
             adapter = viewAdapter
         }
 
+        super.onStart()
     }
 }
