+++ app/src/main/java/com/finnmglas/launcher/extern/Functions.kt
@@ -17,11 +17,10 @@ import android.view.animation.*
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.Toast
-import androidx.core.content.ContextCompat.startActivity
-import com.finnmglas.launcher.ChooseActivity
+import com.finnmglas.launcher.choose.ChooseActivity
 import com.finnmglas.launcher.R
-import com.finnmglas.launcher.SettingsActivity
-import com.finnmglas.launcher.intendedSettingsPause
+import com.finnmglas.launcher.settings.SettingsActivity
+import com.finnmglas.launcher.settings.intendedSettingsPause
 import kotlin.math.roundToInt
