+++ app/src/main/java/com/finnmglas/launcher/settings/SettingsActivity.kt
@@ -1,4 +1,4 @@
-package com.finnmglas.launcher
+package com.finnmglas.launcher.settings
 
 import android.content.Context
 import android.content.Intent
@@ -9,8 +9,8 @@ import android.view.View
 import android.view.WindowManager
 import androidx.appcompat.app.AppCompatActivity
 import androidx.viewpager.widget.ViewPager
+import com.finnmglas.launcher.R
 import com.finnmglas.launcher.extern.*
-import com.finnmglas.launcher.settings.SettingsSectionsPagerAdapter
 import com.google.android.material.tabs.TabLayout
 import kotlinx.android.synthetic.main.activity_settings.*
