+++ app/src/main/java/com/finnmglas/launcher/settings/actions/SettingsFragmentApps.kt
@@ -1,4 +1,4 @@
-package com.finnmglas.launcher.settings
+package com.finnmglas.launcher.settings.actions
 
 import android.content.ActivityNotFoundException
 import android.content.Intent
@@ -10,11 +10,10 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.finnmglas.launcher.ChooseActivity
+import com.finnmglas.launcher.choose.ChooseActivity
 import com.finnmglas.launcher.R
 import com.finnmglas.launcher.extern.*
-import com.finnmglas.launcher.intendedSettingsPause
-import com.finnmglas.launcher.settings.actions.ActionsRecyclerAdapter
+import com.finnmglas.launcher.settings.intendedSettingsPause
 import kotlinx.android.synthetic.main.fragment_settings_apps.*
