+++ app/src/main/res/layout/activity_choose.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="?attr/colorPrimary"
-    tools:context=".ChooseActivity">
+    tools:context=".choose.ChooseActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/activity_choose_app_bar"
