+++ app/src/main/res/layout/activity_tutorial.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:background="?attr/colorPrimary"
     android:onClick="clickAnywhere"
-    tools:context=".TutorialActivity">
+    tools:context=".tutorial.TutorialActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/activity_firststartup_app_bar"
